COM-3415 create composable learneredition
