+++ src/core/components/courses/TraineeTable.vue
@@ -34,7 +34,6 @@
 <script>
 import { computed, ref, toRefs } from 'vue';
 import { useStore } from 'vuex';
-import { useRouter } from 'vue-router';
 import { useQuasar } from 'quasar';
 import useVuelidate from '@vuelidate/core';
 import { required } from '@vuelidate/validators';
@@ -49,7 +48,7 @@ import TraineeEditionModal from '@components/courses/TraineeEditionModal';
 import ResponsiveTable from '@components/table/ResponsiveTable';
 import { NotifyNegative, NotifyWarning, NotifyPositive } from '@components/popup/notify';
 import { frPhoneNumber } from '@helpers/vuelidateCustomVal';
-import { TRAINING_ORGANISATION_MANAGER, VENDOR_ADMIN } from '@data/constants';
+import { useLearnersEdition } from '@composables/learnersEdition';
 
 export default {
   name: 'TraineeTable',
@@ -71,7 +70,8 @@ export default {
 
     const $q = useQuasar();
     const $store = useStore();
-    const $router = useRouter();
+
+    const { canAccesOrEditTrainee } = useLearnersEdition();
 
     const traineePagination = ref({ rowsPerPage: 0, sortBy: 'lastname' });
     const traineeEditionModal = ref(false);
@@ -176,18 +176,7 @@ export default {
       }
     };
 
-    const isVendorInterface = /\/ad\//.test($router.currentRoute.value.path);
-    const loggedUser = computed(() => $store.state.main.loggedUser);
-    const isRofOrAdmin = computed(() => [TRAINING_ORGANISATION_MANAGER, VENDOR_ADMIN]
-      .includes(get(loggedUser.value, 'role.vendor.name')));
-
-    const canEditTrainee = (trainee) => {
-      const loggedUserCompany = get(loggedUser.value, 'company._id');
-      const traineeCurrentCompany = trainee.company;
-
-      return !course.value.archivedAt &&
-        ((isVendorInterface && isRofOrAdmin) || (!isVendorInterface && loggedUserCompany === traineeCurrentCompany));
-    };
+    const canEditTrainee = trainee => canAccesOrEditTrainee(trainee) && !course.value.archivedAt;
 
     return {
       // Data
